decontam: tidy renamed-path docstrings + drop ``--priority production``
Path strings inside ``all_sources_decon.py`` and ``count_docs.py`` still
referenced the old ``experiments/decontamination/...`` location after the
rename in 3b7ad8413. Update the docstring submit commands and one inline
comment to use ``experiments/datakit/decontam/...``. Also drop the
``--priority production`` flag from the all_sources_decon submit command
in the docstring -- production was the V0 default but interactive is
the better fit for follow-up reruns with the AA-aligned eval suite.
